implemented add and delete actions using token and added regex for registration post method

+++ semestrovka/Controllers/AdminController.cs
@@ -1,4 +1,5 @@
 using System.Net;
+using System.Text.RegularExpressions;
 using semestrovka.Attributes;
 using semestrovka.DAOs;
 using semestrovka.DTOs;
@@ -13,12 +14,21 @@ namespace semestrovka.Controllers;
 [Controller("Admin")]
 public class AdminController
 {
+    private readonly HttpListenerContext _context;
+    private readonly AuthService _authService = new();
+    private readonly CookiesHandlerService _cookiesHandlerService = new();
+
+    public AdminController(HttpListenerContext context)
+    {
+        _context = context;
+    }
+
     [Authorize]
     [Get("Index")]
-    public static ResponseMessage Index(HttpListenerContext context)
+    public ResponseMessage Index(HttpListenerContext context)
     {
         var cookie = context.Request.Cookies.First(i => i.Name == "Authorization");
-        var users = new UserDao().GetUsers().Select(i => new { i.Id, token = i.Email + i.Password});
+        var users = new UserDao().GetUsers().Select(i => new { i.Id, token = i.Email + i.Password });
         var userId = users.First(i => i.token == cookie.Value).Id;
         var mappedPage = PageMapper.MapAdminPage(userId);
         return mappedPage != ""
@@ -27,14 +37,14 @@ public class AdminController
     }
 
     [Get("Login")]
-    public static ResponseMessage Login()
+    public ResponseMessage Login()
     {
         var page = StaticFilesHandler.FetchStaticFile("login.html");
         return new ResponseMessage(200, page);
     }
-    
+
     [Post("Login")]
-    public static ResponseMessage Login(string email, string password)
+    public ResponseMessage Login(string email, string password)
     {
         var userToAuthorize = new UserDto
         {
@@ -56,21 +66,26 @@ public class AdminController
     }
 
     [Get("Logout")]
-    public static ResponseMessage Logout()
+    public ResponseMessage Logout()
     {
         return new ResponseMessage(200, "setcookie Authorization _");
     }
-    
+
     [Get("Registration")]
-    public static ResponseMessage Registration()
+    public ResponseMessage Registration()
     {
         var page = StaticFilesHandler.FetchStaticFile("registration.html");
         return new ResponseMessage(200, page);
     }
-    
+
     [Post("Registration")]
-    public static ResponseMessage Registration(string email, string password)
+    public ResponseMessage Registration(string email, string password)
     {
+        var regex = new Regex("([a-zA-Z0-9._-]+@[a-zA-Z0-9._-]+\\.[a-zA-Z0-9_-]+)");
+
+        if (!regex.IsMatch(email))
+            return new ResponseMessage(401, "Wrong email");
+        
         var userToRegister = new UserDto
         {
             Email = email,
@@ -82,32 +97,30 @@ public class AdminController
             ? new ResponseMessage(200, "Successfully registered")
             : new ResponseMessage(404, "Some error occured");
     }
-    
+
     [Authorize]
     [Get("AddMission")]
-    public static ResponseMessage AddMission()
+    public ResponseMessage AddMission()
     {
         var page = StaticFilesHandler.FetchStaticFile("addMission.html");
         return page != ""
             ? new ResponseMessage(200, page)
             : new ResponseMessage(404, "Couldn't get mission adding page");
     }
-    
+
     [Authorize]
     [Post("AddMission")]
-    public static ResponseMessage AddMission(
+    public ResponseMessage AddMission(
         string title,
         string continent,
         string country,
         string city,
         string description,
-        string url,
-        int userId
+        string url
     )
     {
         var missionReceived = new MissionAddingDto
         {
-            UserId = userId,
             Title = title,
             Continent = continent,
             Country = country,
@@ -115,9 +128,17 @@ public class AdminController
             Description = description,
             Url = url
         };
-        
+
         var mission = MissionMapper.MapMissionAddingDtoToMission(missionReceived);
         var missionDao = new MissionDao();
+        var userIdTuple = _cookiesHandlerService
+            .FetchTokenFromCookiesAndGetUserId(_context, _authService, new UserDao());
+        
+        if (!userIdTuple.Item1)
+            return new ResponseMessage(404, "Something went wrong");
+
+        mission.UserId = userIdTuple.Item2;
+        
         return missionDao.CreateMission(mission)
             ? new ResponseMessage(200, "Mission was successfully added")
             : new ResponseMessage(404, "Something went wrong");
@@ -125,16 +146,24 @@ public class AdminController
 
     [Authorize]
     [Post("DeleteMission")]
-    public static ResponseMessage DeleteMission(int id)
+    public ResponseMessage DeleteMission(int id)
     {
         var missionDao = new MissionDao();
+        
+        var tokenTuple = _cookiesHandlerService.FetchTokenFromCookies(_context);
+        if (!tokenTuple.Item1)
+            return new ResponseMessage(404, "Something went wrong");
+        
+        if (!_authService.ValidateToken(new UserDao(), tokenTuple.Item2))
+            return new ResponseMessage(403, "You do not have access for that");
+        
         var mission = missionDao.DeleteMission(id);
         return new ResponseMessage(200, mission?.Title ?? "");
     }
 
     [Authorize]
     [Get("AddScreenshot")]
-    public static ResponseMessage AddScreenshot()
+    public ResponseMessage AddScreenshot()
     {
         var page = StaticFilesHandler.FetchStaticFile("addScreenshot.html");
         return page != ""
@@ -144,21 +173,27 @@ public class AdminController
 
     [Authorize]
     [Post("AddScreenshot")]
-    public static ResponseMessage AddScreenshot(
+    public ResponseMessage AddScreenshot(
         string text,
-        string url,
-        int userId)
+        string url)
     {
         var screenshotReceived = new ScreenshotCardAddingDto
         {
             Text = text,
-            Url = url,
-            UserId = userId
+            Url = url
         };
 
         var screenshotCard = ScreenshotCardMapper
             .MapScreenshotCardAddingDtoToScreenshotCard(screenshotReceived);
         var screenshotCardDao = new ScreenshotCardDao();
+        var userIdTuple = _cookiesHandlerService
+            .FetchTokenFromCookiesAndGetUserId(_context, _authService, new UserDao());
+        
+        if (!userIdTuple.Item1)
+            return new ResponseMessage(404, "Something went wrong");
+
+        screenshotCard.UserId = userIdTuple.Item2;
+        
         return screenshotCardDao.CreateScreenshotCard(screenshotCard)
             ? new ResponseMessage(200, "Screenshot was successfully added")
             : new ResponseMessage(404, "Something went wrong");
@@ -166,16 +201,24 @@ public class AdminController
 
     [Authorize]
     [Post("DeleteScreenshot")]
-    public static ResponseMessage DeleteScreenshot(int id)
+    public ResponseMessage DeleteScreenshot(int id)
     {
         var screenshotCardDao = new ScreenshotCardDao();
+        
+        var tokenTuple = _cookiesHandlerService.FetchTokenFromCookies(_context);
+        if (!tokenTuple.Item1)
+            return new ResponseMessage(404, "Something went wrong");
+        
+        if (!_authService.ValidateToken(new UserDao(), tokenTuple.Item2))
+            return new ResponseMessage(403, "You do not have access for that");
+        
         var screenshotCard = screenshotCardDao.DeleteScreenshotCard(id);
         return new ResponseMessage(200, screenshotCard?.Text ?? "");
     }
-    
+
     [Authorize]
     [Get("AddVideo")]
-    public static ResponseMessage AddVideo()
+    public ResponseMessage AddVideo()
     {
         var page = StaticFilesHandler.FetchStaticFile("addVideo.html");
         return page != ""
@@ -185,22 +228,29 @@ public class AdminController
 
     [Authorize]
     [Post("AddVideo")]
-    public static ResponseMessage AddVideo(
+    public ResponseMessage AddVideo(
         string text,
         string preview,
-        string redirect,
-        int userId)
+        string redirect)
     {
         var receivedVideo = new VideoCardAddingDto
         {
             Text = text,
             PreviewUrl = preview,
-            RedirectUrl = redirect,
-            UserId = userId
+            RedirectUrl = redirect
         };
 
         var videoCard = VideoCardMapper.MapVideoCardAddingDtoToVideoCard(receivedVideo);
         var videoCardDao = new VideoCardDao();
+        
+        var userIdTuple = _cookiesHandlerService
+            .FetchTokenFromCookiesAndGetUserId(_context, _authService, new UserDao());
+        
+        if (!userIdTuple.Item1)
+            return new ResponseMessage(404, "Something went wrong");
+
+        videoCard.UserId = userIdTuple.Item2;
+        
         return videoCardDao.CreateVideoCard(videoCard)
             ? new ResponseMessage(200, "Screenshot was successfully added")
             : new ResponseMessage(404, "Something went wrong");
@@ -208,9 +258,17 @@ public class AdminController
 
     [Authorize]
     [Post("DeleteVideo")]
-    public static ResponseMessage DeleteVideo(int id)
+    public ResponseMessage DeleteVideo(int id)
     {
         var videoCardDao = new VideoCardDao();
+
+        var tokenTuple = _cookiesHandlerService.FetchTokenFromCookies(_context);
+        if (!tokenTuple.Item1)
+            return new ResponseMessage(404, "Something went wrong");
+        
+        if (!_authService.ValidateToken(new UserDao(), tokenTuple.Item2))
+            return new ResponseMessage(403, "You do not have access for that");
+        
         var videoCard = videoCardDao.DeleteVideoCard(id);
         return new ResponseMessage(200, videoCard?.Text ?? "");
     }
